requirements.txt
fastapi
uvicorn[standard]
python-dotenv
langchain
langchain-community
chromadb
llama-cpp-python
requests
duckduckgo-search
python-multipart

# Dependências para Fine-Tuning (Fase 2)
transformers
datasets
peft
bitsandbytes
trl
accelerate

# Orquestração multiagente
autogen-agentchat
tqdm
huggingface_hub
